Refactor: 네비게이션 Elizabath가 말한대로 더 컴포넌트화 조업과 관제 두 부분에서 이제 네비게이션바가 다르게 뜨고, 서로 다르게 수정가능
